+++ src/pipecatcloud/cli/entry_point.py
@@ -15,6 +15,7 @@ from pipecatcloud.cli.commands.auth import auth_cli
 from pipecatcloud.cli.commands.deploy import create_deploy_command
 from pipecatcloud.cli.commands.docker import create_docker_command
 from pipecatcloud.cli.commands.organizations import organization_cli
+from pipecatcloud.cli.commands.regions import regions_cli
 from pipecatcloud.cli.commands.secrets import secrets_cli
 from pipecatcloud.cli.config import config
 from pipecatcloud.exception import ConfigFileError
@@ -78,6 +79,7 @@ create_deploy_command(entrypoint_cli_typer)
 create_docker_command(entrypoint_cli_typer)
 entrypoint_cli_typer.add_typer(auth_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(organization_cli, rich_help_panel="Commands")
+entrypoint_cli_typer.add_typer(regions_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(secrets_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(agent_cli, rich_help_panel="Commands")